+++ mintedge/services.py
@@ -2,7 +2,7 @@ class Service:
     def __init__(
         self,
         name: str,
-        workload: float,
+        workload: int,
         arrival_rate: float,
         input_size: int,
         output_size: int,
@@ -13,7 +13,7 @@ class Service:
         station redirecting traffic to the edge server.
         Args:
             name (str): Name of the service
-            workload (float): Number of ops necessary to complete the request
+            workload (int): Number of ops necessary to complete the request
             arrival_rate (float): Number of req per sec caused by this service
             input_size (int): Bytes of data per request made to this service
             output_size (int): Bytes of data per request returned by this service
@@ -24,7 +24,7 @@ class Service:
         self.arrival_rate = arrival_rate
         self.input_size = input_size * 8  # bits
         self.output_size = output_size * 8  # bits
-        self.max_delay = max_delay
+        self.max_delay = max_delay  # seconds
 
     def __repr__(self):
         return self.name
